pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-user-story-development"
version = "1.0.0"
description = "Claude Code plugin for multi-agent ETL development from user stories"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Program Unify", email = "unify@emstas.com"}
]
keywords = ["claude", "pyspark", "etl", "azure", "devops", "multi-agent", "orchestration", "user-stories"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "anthropic>=0.7.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
    "azure-devops>=7.0.0",
    "azure-identity>=1.13.0",
    "azure-storage-blob>=12.17.0",
    "azure-keyvault-secrets>=4.7.0",
    "mcp>=0.1.0",
]

[project.optional-dependencies]
pyspark = [
    "pyspark>=3.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
all = [
    "pyspark>=3.4.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
multi-agent-scaffold = "multi_agent_user_story_development.orchestrator:main"
multi-agent-mcp = "multi_agent_user_story_development.mcp.server:main"

[project.urls]
Repository = "https://github.com/emstas/multi-agent-user-story-development"
Issues = "https://github.com/emstas/multi-agent-user-story-development/issues"
Documentation = "https://github.com/emstas/multi-agent-user-story-development/blob/main/README.md"

[tool.setuptools]
packages = ["multi_agent_user_story_development"]

[tool.setuptools.package-data]
multi_agent_user_story_development = ["templates/*.md", "prompts/*.py"]

[tool.ruff]
line-length = 240
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 240
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=multi_agent_user_story_development --cov-report=term-missing"
